requirements.txt
ultralytics
opencv-python
torch
torchvision
roboflow
python-dotenv